Refactor Form, move storage verify functions and split BuildSection into components
