+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/TeamSection/TeamMemberProjectsView.js
@@ -7,11 +7,6 @@ import { Column, Line } from '../../../../UI/Grid';
 import EmptyMessage from '../../../../UI/EmptyMessage';
 import AlertMessage from '../../../../UI/AlertMessage';
 import { List } from '../../../../UI/List';
-import {
-  ProjectFileListItem,
-  getProjectLineHeight,
-  transformCloudProjectsIntoFileMetadataWithStorageProviderName,
-} from '../BuildSection';
 import { useResponsiveWindowWidth } from '../../../../UI/Reponsive/ResponsiveWindowMeasurer';
 import { type FileMetadataAndStorageProviderName } from '../../../../ProjectsStorage';
 import { type StorageProvider } from '../../../../ProjectsStorage';
@@ -22,6 +17,11 @@ import FlatButton from '../../../../UI/FlatButton';
 import Skeleton from '@material-ui/lab/Skeleton';
 import ListItem from '@material-ui/core/ListItem';
 import Text from '../../../../UI/Text';
+import {
+  getProjectLineHeight,
+  transformCloudProjectsIntoFileMetadataWithStorageProviderName,
+} from '../BuildSection/utils';
+import ProjectFileListItem from '../BuildSection/ProjectFileListItem';
 
 const styles = {
   listItem: {
