+++ newIDE/app/src/Profile/ChangeEmailDialog.js
@@ -14,6 +14,7 @@ import { ColumnStackLayout } from '../UI/Layout';
 import TextField from '../UI/TextField';
 import { getEmailErrorText } from './CreateAccountDialog';
 import { emailRegex } from './ForgotPasswordDialog';
+import Form from '../UI/Form';
 
 type Props = {|
   firebaseUser: FirebaseUser,
@@ -71,15 +72,7 @@ const ChangeEmailDialog = ({
       onApply={doChangeEmail}
       open
     >
-      <form
-        onSubmit={event => {
-          // Prevent browser to navigate on form submission.
-          event.preventDefault();
-          doChangeEmail();
-        }}
-        autoComplete="on"
-        name="changeEmail"
-      >
+      <Form onSubmit={doChangeEmail} autoComplete="on" name="changeEmail">
         <ColumnStackLayout noMargin>
           <TextField
             value={email}
@@ -102,7 +95,7 @@ const ChangeEmailDialog = ({
             }}
           />
         </ColumnStackLayout>
-      </form>
+      </Form>
     </Dialog>
   );
 };
