+++ newIDE/app/src/Profile/ForgotPasswordDialog.js
@@ -9,6 +9,7 @@ import TextField from '../UI/TextField';
 import { type ForgotPasswordForm } from '../Utils/GDevelopServices/Authentication';
 import LeftLoader from '../UI/LeftLoader';
 import Text from '../UI/Text';
+import Form from '../UI/Form';
 
 export const emailRegex = /^(.+)@(.+)$/;
 
@@ -77,15 +78,7 @@ const ForgotPasswordDialog = ({ onClose, onForgotPassword }: Props) => {
           </Text>
         </Column>
       ) : (
-        <form
-          onSubmit={event => {
-            // Prevent browser to navigate on form submission.
-            event.preventDefault();
-            doResetPassword();
-          }}
-          autoComplete="on"
-          name="resetPassword"
-        >
+        <Form onSubmit={doResetPassword} autoComplete="on" name="resetPassword">
           <Column noMargin>
             <TextField
               autoFocus="desktop"
@@ -110,7 +103,7 @@ const ForgotPasswordDialog = ({ onClose, onForgotPassword }: Props) => {
               }}
             />
           </Column>
-        </form>
+        </Form>
       )}
     </Dialog>
   );
