+++ newIDE/app/src/UI/Alert/AlertContext.js
@@ -12,6 +12,7 @@ export type ShowAlertDialogOptionsWithCallback = {|
   ...ShowAlertDialogOptions,
   callback: Function,
 |};
+export type ShowAlertFunction = ShowAlertDialogOptions => Promise<void>;
 
 // Confirm
 export type ShowConfirmDialogOptions = {|
@@ -26,6 +27,7 @@ export type ShowConfirmDialogOptionsWithCallback = {|
   ...ShowConfirmDialogOptions,
   callback: Function,
 |};
+export type ShowConfirmFunction = ShowConfirmDialogOptions => Promise<boolean>;
 
 // Confirm Delete
 export type ShowConfirmDeleteDialogOptions = {|
@@ -40,6 +42,7 @@ export type ShowConfirmDeleteDialogOptionsWithCallback = {|
   ...ShowConfirmDeleteDialogOptions,
   callback: Function,
 |};
+export type ShowConfirmDeleteFunction = ShowConfirmDeleteDialogOptions => Promise<boolean>;
 
 export type ConfirmState = {|
   showAlertDialog: ShowAlertDialogOptionsWithCallback => void,
