+++ newIDE/app/src/UI/Form.js
@@ -1,0 +1,37 @@
+// @flow
+import * as React from 'react';
+
+type Props = {|
+  onSubmit: () => void | Promise<void>,
+  autoComplete?: 'on' | 'off',
+  name: string,
+  children: React.Node,
+|};
+
+const Form = ({
+  onSubmit,
+  autoComplete = 'off', // Default to 'off' to avoid browser autofill.
+  name,
+  children,
+}: Props) => {
+  return (
+    <form
+      onSubmit={event => {
+        // Prevent browser to navigate on form submission.
+        event.preventDefault();
+        onSubmit();
+      }}
+      autoComplete={autoComplete}
+      name={name}
+    >
+      {children}
+      {/*
+        This input is needed so that the browser submits the form when
+        Enter key is pressed. See https://stackoverflow.com/questions/4196681/form-not-submitting-when-pressing-enter
+      */}
+      <input type="submit" value="Submit" style={{ display: 'none' }} />
+    </form>
+  );
+};
+
+export default Form;
